requirements.txt
django
djangorestframework
django-cors-headers
qrcode
pillow
pyjwt
coverage
django-heroku
gunicorn
django-environ
python-dateutil
requests
google-auth
